termék listázás működik a lokális adatbázisból, működik a bejelentkezés és a regisztráció

+++ app/src/main/java/hu/unideb/shoppinglist/database/model/Product.kt
@@ -12,5 +12,5 @@ data class Product(
 ) {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    var Id: Int? = null
+    var Id: Long? = null
 }
